requirements.txt
# Core Dependencies
opencv-python>=4.8.0,<5.0.0
Pillow>=10.0.0,<13.0.0
numpy>=1.24.0,<3.0.0

# Optional: For advanced features
# pandas>=2.0.0  # For log analysis
# matplotlib>=3.7.0  # For visualization
# requests>=2.31.0  # For webhook integration
# python-dotenv>=1.0.0  # For environment variables